Suppress python{3,}-gi → python{3,}-cairo dependency
Since we moved to pybuild, we pick up a dependency from python{3,}-gi to
python{3,}-cairo, because it is in requirements.txt.

Add a py3dist-overrides file to map this to the empty package, so that
dh_python{3,} doesn't generate the dependency for us - it is unnecessary
in python{3,}-gi and python{3,}-gi-cairo, which needs it, has it
specified manually.

Closes: #909703